Create a symlink for -latest as seed stage
Create a symlink for the latest name from the available stages3 always,
that is then used as the seed for building the new stages. To ensure
creation, always copy instead of move the stages to the destination
afterwards. In addition to this, we need a cleanup script, as before
only the isos and stages3 were moved, leaving the stage[12] eating up
space behind.

Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>
Proposed-by: Jorge Manuel B. S. Vicetto <jmbsvicetto@gentoo.org>
Acked-by: Andrew Gaffney <agaffney@gentoo.org>

svn path=/trunk/; revision=769