tools-hardened/desktop/fluxbox-run.sh: Fixes fluxbox config file issue.
The Fluxbox menu can either be a standard default or a custom file. While
there is already a custom menu file in existance under files/usermenu, it
was not previously being copied over prior to the creation of the ISO. This
caused the user to not have a custom menu file upon boot. To fix this, a line
has been added to the script which creates the .fluxbox directory prior to
building the image.