tools-hardened/desktop: Adds ability to set wallpaper for fluxbox.
Previously, if you were to boot this image you would be greeted with
an error message telling you that the background could not be set.
To fix this the package eterm was added to the fluxbox world file
to be pulled in on build-time. However, this only solved the problem
of being unable to set the background manually. To have fluxbox set
it automatically the .xinit was altered to call "exec startfluxbox".
When fluxbox is called by that command it searches for and executes
the startup file located in ~/.fluxbox/. This commit also adds a
custom startup file which sets the background to the default background:
/usr/share/backgrounds/background.jpg. This commit also includes one
unrelated alteration to fluxbox-run.sh which changes the fluxbox init
file to look at the custom menu in /usr/share/fluxbox/usermenu instead
of ~/.fluxbox/my-menu.