catalyst-auto: add a time delay option
We usually run catalyst on systems via cronjob, and do so at weekly
(or so) intervals.  But if a failure hits a run, it will be a while
before we know if changes fixed the situation.  Instead, let's push
the delay logic into the script itself so we can run this tool once
a day.  When it passes, it will take care of quitting early.  When
it fails, it will retry once a day until it passes.