remove unneeded/obsolete gk_mainargs from *cd-stage2*.spec
There is no reason for the livecd initramfs to need to be able to open
mdraid/dmraid/lvm arrays, so we can remove this support from the
genkernel call.  The livecds will still have all the tools needed to
create/maintain/use these kinds of volumes later.

Modern genkernel also autodetects makeopts so we do not need to set that
explicitly.