specs: add /boot to livecd/empty to work around bug 661084
Due to bug 661084, the /boot/System* and other wildcard matches in
livecd/rm are not functional.  This causes us to ship a duplicate copy
of kernel/initramfs, wasting space.

I've added /boot to livecd/empty to work around this for now.  I only
added this to specs where /boot/* things were already in livecd/rm.