portage/isos/savedconfig: savedconfig for linux-firmware
This adds a new savedconfig file for linux-firmware to cut out firmware
types that we obviously don't need on installcds, such as multimedia/dvb
cards.  This file is unversioned so it'll apply to future linux-firmware
releases, so it may need minor tweaks over time as new unwanted
firmwares are added upstream.

This file doesn't take effect unless/until we set USE=savedconfig on
linux-firmware, this hasn't been done yet.