specs: Enable USE=compile-locales for installcd-stage1
This patch causes glibc (starting with version 2.30) to be remerged in
installcd-stage1 with USE=compile-locales.

Available with glibc-2.30, USE=compile-locales will generate the locales
in src_install() and include them in the binary package.

This enables catalyst to reduce the locales produced in stages 1-3 to a
minimal set without causing the ISOs to have a limited set of locales.

Bug: https://bugs.gentoo.org/710762
Signed-off-by: Matt Turner <mattst88@gentoo.org>