portage/isos/package.use: enable symlink for gentoo-sources
kernel-2.eclass was changed so that the /usr/src/linux symlink is no
longer created by default, even when it doesn't exist at all,unless
USE=symlink is explicitly set.  This breaks an assumption we make in
catalyst.  Explicitly setting symlink will work around this issue until
we fix it up in catalyst.

Signed-off-by: Ben Kohler <bkohler@gentoo.org>