specs: Drop USE=static-libs from installcd
This was added in 2011 in commit 1d973914 ("Add static-libs to USE for
the installcd-stage1 specs to fix the dev-libs/libgcrypt failure
build.") to resolve an unreferenced problem.

Reduces the number of packages built for the alpha installcd from 148 to
105, because presumably 43 packages that are already part of the stage3
seed were being rebuilt with USE=static-libs enabled. This is pointless
because the livecd-stage2 spec's livecd/rm contains /lib*/*.a and
/usr/lib*/*.a, so all the static libraries are being deleted anyway.

Signed-off-by: Matt Turner <mattst88@gentoo.org>