Merge live media package lists for amd64 & x86
Using the base amd64 admin as the template as it's the most up to date.

Reasoning behind this is CDROM limits of the past are no longer limiting factor so we can go over the 700MB. The other benefit is the reduction in builds were making and supporting a week which should hopefully allow us to provide more builds in the future like an arm64 livegui.

Signed-off-by: Ian Jordan <immoloism@gmail.com>