diskimage-qemu: mask latest glibc
Copied over the p.mask for the latest version of glibc which
is known to cause issue with qemu builds for now.

Also copied the dev-lang/rust mask to use rust-bin to save time.
Not needed yet, but I'm sure the day will come so we are ready for it.

Bug: https://bugs.gentoo.org/961307
Signed-off-by: Ian Jordan <immoloism@gmail.com>
Signed-off-by: Andreas K. Hüttel <dilfridge@gentoo.org>