releases/specs/{amd64,arm64,x86}: enable support for booting with NTFS filesystem Add "dmsquash-live-ntfs" Dracut module which enable the use of booting an installation media with NTFS.
The UEFI spec encompasses the use of FAT, but UEFI is happy to boot from any filesystem as long as the firmware as a proper UEFI driver for it.
This is especially relevant because FAT has some design limitations being it very old, such as not allowing >4 GiB files.

Other limitations are:
- label
  - can be no longer than 11 characters
  - must use uppercase letters
  - characters below 0x20 are not allowed
  - characters *?.,;:/\|+=<>[]" are not allowed
  - can't start with a space character
  - must contain English characters only
- do not support symbolic links because their design predates the feature and lacks a mechanism to mark files as links
- do not support UNIX file permissions

Now, because Windows ISO of the last few years contains at least one file that is larger than 4 GiB, Microsoft put pressure on OEM to make sure they include an UEFI NTFS driver in their firmware.
This is true for recent (usually from last 10 years) motherboards from major brands (Gigabyte, ASUS, MSI, Intel, etc.)

This is because, by design, Microsoft creates Windows ISO that are FST (File System Transposition) compatible, meaning that the end-user is able to take a USB drive, format to NTFS, copy the content of the ISO into it and boot it in UEFI without any additional step for the user (unlike other Linux distros which rely on ISOHybrid, so the user is forced to use an additional tool like dd or Rufus to make a bootable drive).

What this PR is all about is adding support for NTFS filesystem, such that Gentoo install media can now be bootable from this fs, just like Windows does.

It would support both fs by having the initramfs include FAT32 and NTFS modules.
The actual module which is used at runtime depends on how the user chooses to format the media the ISO is written to (FAT or NTFS).

Because this change does increase RAM usage a bit, this PR only enable this feature for known systems architectures (x86,x86_64,arm64)